+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Changed
+- Moved Phenomodels code under a dedicated blueprint
 ### Added
 - A summary table of pinned variants on the cancer case general report
 - New openable matching causatives and managed variants lists for default gene panels only for convenience